fix(party parsing): Fixes very dumb typo.
Looks like I mistyped something and didn't catch it because it replaced a single character that
isn't caught in our testing. Why isn't it caught in testing? Because it only happens when you have
multiple tasks running simultaneously, which we don't have an easy tool for doing in tests. Oy.