fix(party parsing): Only preserve attorneys on *terminated* parties.
Otherwise, if a party is not terminated and it has an extraneous attorney, that attorney gets
wrongly preserved.

This improves things, but we still have the issue of a terminated party with an extraneous attorney.
In that case, the extraneous attorney is preserved because we assume that preserving attorneys on
terminated parties makes sense. Not much more we can do here, I'm afraid.