fix(pacer reprocessing script): Catches XMLSyntaxError
These seem to happen when the local copy of the XML is an empty file. I fixed a few of these
manually, but that's very slow and it's better to just catch it and move on. We'll see how big the
issue is this way, and then we can choose how to address it.

Related to: #800