feat(RECAP APIs): Ensures correct courts are used with correct upload types
This changes the API slightly when it is used incorrectly. Instead of allowing you to upload a
an appellate docket to the district docket endpoint or vice versa, this throws a 400 error instead,
and includes a useful message: "Didn't you mean to..."

Should keep our data a bit safer on the off chance that an uploaded docket can somehow be parsed by
one or the other parser.