feat(GDPR): Adds a take out feature (sort of)
All this does for now is sends a request to the admins. At that point, it'll be up to them (me) to
generate the exported file and get it to the user. Hopefully nobody will ever press this button. If
numerous people do, I'll make it into a proper feature.