fix(free opinion scraper): Ensures that we always quit...eventually.
I was seeing in production that sometimes this script would never finish. Really not sure how/why
this was happening, but looking at the log I'd see something like:

    INFO:cl.lib.command_utils:Court wawd still in progress. Delaying at least 30s.
    ...A million more lines like this...
    INFO:cl.lib.command_utils:Court wawd still in progress. Delaying at least 30s.

It varied which court was the culprit, and it was always a LONG log of these lines. The fix in this
commit should make it so that we try about 20 times (which might take around ten minutes), and if
we don't have a good result by then, we move on.