fix(fjc import script): Fix quotation handling
I don't know if this is normal or not, but the FJC data sources of some pretty funky values. One
think they do, which we didn't know before now is have values like:

    M/V ""TITANIC""

Which stands for:

    M/V "TITANIC"

(M/V stands for Motor Vessel)

Previously, we had two bugs. First, we didn't remember to strip the quotes off of this. Second, we
didn't notice the weird double-double quote thing ("") that is being done, and so we didn't fix that
either.

This commit fixes both of those errors.