feat(registration flow): Shows the email that signed up and provides a link back
Showing the email is important because sometimes people have typos in their email address when they
sign up.

Providing a link back is a nice-to-have so people can get back to where they were.

Also improves the page responsiveness for different screensizes.

User auth and auth flow is hard to get right.