fix(alert HTML sidebar): Removes an errant if statement
This checks that the user has some kind of GET query. You won't see search results unless you do, so
this doesn't do anything. In an earlier day, when there was no query, instead of showing the
homepage, we showed search results, kind of like a Facebook feed, but those days were a long time
ago.