DocketEntry needs a date_entered field.
Because sometimes we get date_entered instead of date_filed.
They are, of course, not the same. Sometimes.

This should be a datetime, even though it's very hard for us to
actually find the time of this field. Unlike the time of date_entered,
which should also be a datetime and is easy to find (at least
sometimes), however that requires a flag day so it's a change for
another day.