fix(recap alert): Ensures recap alerts remain hidden (see note re git)
So, if you look at the git history from today, you'll see that it's a huge mess. I had a feature
branch for recap search alerts, but I accidentally put code in it for alert enable/delete. I then
tried to move those commits over to a new feature branch, somehow detached them from the world,
and then had to cherry pick them back into place after putting them on a temp branch.

The cherry picking went poorly since I was trying to skip the earlier commits on that feature
branch, which, of course, the code relied on. But eventually my enable/disable code was cherry
picked. But then I realized it didn't work and tried to cherry pick earlier commits. That made the
order of commits get all messed up, and, well, getting it right has been horrible.

Anyway, one small mistake in git (forgetting to switch branches before starting a different
feature), has resulted in much frustration and ire towards git --- and not for the first time, I
might add.

Setting all of that aside, RECAP search is not yet ready, so this commit makes sure it's hidden.
Normally it would be hidden already, except that all this terrible stuff with git happened, and
part of that was that it got free. This commit fixes that, and hides it again.