fix(search api): Fixes an off-by-one that affected every page except page 1.
The first page would have 20 results, but every page after that would have 19. Looks like I had the
wrong idea in my comments as well as in the code, so I fixed both.

I attempted for far too long today to add a test for this code, but I couldn't find an easy way to
do so. Either I needed to create a big test fixture with 40+ items in it (because page size is 20),
and using that, I needed to test the two queries. Or I needed to do a lower-level test, which was
difficult to get right since the arguments to the get_object_list and the SolrList objects are
tricky to create and test. In the end, I just tested this manually, which isn't great, but hopefully
we don't change the code here much, so it'll have to be good enough, at least for now.

This also tweaks things a bit so that we use get_page_size, which is a better way to get the page
size since it knows how to look in the Django settings and in the class settings.