fix(recap search): Docket entries order properly when single docket result
This little tweak is important because when people filter to only one docket (which the UI makes
easy), they previously lost their ability to sort. With this change, the docket entries get
sorted too when you're in a docket query. This way, you can see the most recent 500 items, or the
most relevant (or the oldest, if you like).

By default, docket entries are still sorted by relevancy.

There is on tweak here that I don't like. In the template, we remove the line that previously
sorted the five most relevant items by docket entry number. So now they'll be in true relevancy
order, which is fine, but might confuse some folks.