feat(docket alerts): New task and new task enqueueing helper for docket alerts.
The big picture here is that the helper method checks for a sentinel key. If the key is present, it
does nothing. If the key is missing, it creates the key and it enqueues a new task to send the
emails. Note that the enqueued task is supposed to delete the sentinel key, but if the task fails,
the sentinel is configured with a timer to ensure that it expires eventually anyway.

The task itself is pretty straightforward. Basically, we send an email to anybody that is subscribed
to the docket. In the email, we list any new docket entries that were created since our timer began.
Once the emails go out, we remove the sentinel value.