fix(rss/alerts tasks): Moves alert sending to be outside the transaction block.
I launched the alert feature today, and I am getting a lot of emails with the error below from
Celery:

    Task was called with args: (7170457, datetime.datetime(2018, 6, 14, 23, 36, 41, 449891, tzinfo=<UTC>)) kwargs: {}.

    The contents of the full traceback was:

    Traceback (most recent call last):
      File "/celery/app/trace.py", line 240, in trace_task
        R = retval = fun(*args, **kwargs)
      File "/celery/app/trace.py", line 438, in __protected_call__
        return self.run(*args, **kwargs)
      File "/var/www/courtlistener/cl/alerts/tasks.py", line 55, in send_docket_alert
        docket = Docket.objects.get(pk=d_pk)
      File "/django/db/models/manager.py", line 127, in manager_method
        return getattr(self.get_queryset(), name)(*args, **kwargs)
      File "/django/db/models/query.py", line 334, in get
        self.model._meta.object_name
    DoesNotExist: Docket matching query does not exist.

Basically, the docket didn't exist yet. But this is weird. If the docket doesn't exist, how did we
pass it's PK to the enqueue function? Since we did that, it must have gotten deleted. But we don't
do that. So something weird must be happening.

The answer, I think, is that the transaction block is sorta creating the docket, but when we create
a separate celery task that accesses that docket, that SELECT isn't a part of the same transaction.
Thus, it's possible that that SELECT will happen before the transaction is closed (and the docket is
*actually* created). If that happens, from the other task's perspective, the docket doesn't exist
yet, and an error is thrown.

So: Only enqueue once we've exited the transaction. Should be simple enough!