fix(docket alerts): Makes it so GET requests fail gracefully.
Not sure what's happening. It might be old user agents, it might be crawlers, but somehow we're
getting lots of this error:

    int() argument must be a string or a number, not 'SimpleLazyObject'

All of the requests are GET requests, which means that somebody/something clicked on the link and
the JS didn't work properly. This makes sure that if somebody does a GET, then we just knock them
out of the place.