feat(user welcome emails): Forces separate threads in gmail interface via the X-Entity-Ref-ID header
This is a convenience/safety for admins like me. The problem is that a lot of people respond to the
welcome emails, but all of their responses go into one massive gmail thread, where it's very easy
to lose them. According to...

    https://stackoverflow.com/a/25435722/64911

...it's possible to create separate threads in some email clients by including the X-Entity-Ref-ID
header. Alas, this header is entirely undocumented, so I add it here, and we'll see if Stackoverflow
has good information or not in this instance. Watch for a revert if this doesn't work!