feat(user login): Provides a better error for folks trying to log in with their email address
We just heard from a user that was struggling to log in. The problem? He was using his email address
instead of his username. One fix would be to allow both, but that's complicated. This is an easier
way to at least provide a tip of what the problem may be.