fix(court regexes): Cleans up the regexes and adds a new method for searching them.
Wow, it's been a while since I've peeked in here, and man this isn't a beautiful part of the code.
The biggest thing I'm noticing is that some of the regexes are hilariously broad. This isn't the
end of the world, provided that they're applied in the same order, and that they don't conflict. To
address this, I ensured that we apply the regexes in the correct order, from most specific to least,
and I identified what I think are all of the conflicting ones (it appears to only happen between
district and bankruptcy courts). When there is a potential conflict, an assertion is raised.

The following commits will bring the rest of the code up to speed with the changes here. Happily,
none of our production code actually uses this. Yet.