feat(configs): Moves REDIS and Solr configurations to the private settings file (REQUIRES CHANGES)
This moves a few settings to the private settings file that were previously in the public settings
file. If you simply deploy this code, you will *not* have a functioning system. Instead, you must
add three new settings to your private settings file, as are exemplified in the 05-private.example
file:

    SOLR_URL = 'http://127.0.0.1:8983'

    REDIS_HOST = 'localhost'
    REDIS_PORT = 6379

If you add those settings and then deploy this code, you should be good to go.