fix(docket html): Moves the appeal from info into the OG court section.
This was, unfortunately, necessary because for many dockets, we only have the appeal_from_str and
docket_number values. When this was the case, we had a nice header for OG info, but no danged info
under it. This change moves the appeal from info into the OG info section and out of its title.