fix(PACER download tasks): Switches from passing PacerSession objects as args to passing cookiejars
This change is to address the problem identified here:

https://gist.github.com/mlissner/7227c5c1091b3ef61fa34e86006ef6d1

Basically, for some reason when we pass PacerSessions to Celery tasks, they throw AttributeErrors
some of the time. The errors are not reproducible and reviewing the code I can't see how they could
possibly happen. The best theory I've heard is that it has something to do with serialization, so
this changes moves it so we aren't serializing a PacerSession object anymore and we are instead
passing the cookie from the session. This is a glorified Dict, so it should serialize neatly, and
with any luck our AttributeErrors will go away.