fix(ia tasks): Renames the IA session object to avoid conflicts
This may in fact be the culprit to the problem descibed in the previous commit. What happens when
Celery restarts a worker and it grabs what it thinks is in the session variable inside one of the
functions? That variable could be *either* the ia session or the parameter passed by the caller. I
don't know how this kind of conflict is resolved, and frankly, I don't want to know. I *do* know how
to avoid it, and that's what this commit does by renaming the variable.

In other news, establishing an IA session like this, in the module itself, is probably quite bad.