refactor(pacer import tasks): Factor out the code for getting PDFs and saving them to the DB.
This is a fairly substantial refactor to the code that gets and saves PDFs. Previously we had a lot
of duplication, which was fine because at the time it wasn't clear how the code would mature. At
this point it's pretty clear that the way to do this will be a celery chain where the first task
does the downloading and saving, and then later tasks do OCR and Solr indexing.

So what these changes do is factor out two common code paths into their own functions. The first is
to factor out downloading a PDF, and the second is to factor out saving it properly. (Saving is
fairly complex because we do a lot of other things like compute page count, do a hash, mark it as
free, etc.) For a client job I will be downloading more PDFs so having these functions available is
going to be useful.

For the most part these are API compatible with what we had before. The only real change is that we
no longer pass a flag to the get_and_process_pdf function indicating whether it should save the item
to Solr. Instead, that gets migrated to another task in the celery chain. Everything else should
work and look like it did before.