fix(api authentication): Changes the priority of the REST auth classes.
This change was needed in response to freelawproject/recap#252, which identified that when users
were logged into CourtListener, the Chrome extension was sending their cookies and causing things
to fail. The failure was because:

    {"detail":"CSRF Failed: Referer checking failed - no Referer."}

Because the Chrome extension was sending cookies along with its POSTs, but not sending Referer
values. The cookies would be verified by the SessionAuthentication class, and then we'd get the
error.

The change here is to bump up the TokenAuthentication class so it has an opportunity to work before
the session auth fails. According to the DRF docs,

> The authentication schemes are always defined as a list of classes. REST framework will attempt to
> authenticate with each class in the list, and will set request.user and request.auth using the
> return value of the first class that successfully authenticates.

So hopefully by changing the order of this list, Token auth will work, session won't be run, and
we'll all go home happy. Hopefully.

This is hard to test locally and should not impact API users, so I'm going to give this a try live.