fix(solr tasks & scripts): Ensures that if a solr-modifying celery task is created, it can find Solr
The problem we had previously was that the Solr URL was being passed in as an argument to the task.
That's fine in a non-distributed environment, but when you don't know where the task will be run,
you can't know where Solr will *be*. Therefore, you need to be able to look up the Solr URL within
the task itself, not as a parameter.