fix(opinion downloader): Provides longer soft timeout and retries for SoftTimeLimitExceeded excs
We've had some issues with some courts not getting completed properly. It seems to be correlated
with the largest courts making it a doubly bad problem. I think the issue here is that these courts
just have a lot of content and their reports are very slow. Doubling the timeout and allowing
retries should help.