feat(client work): Adds option for doing only student docket numbers
This was necessary because so many of the student docket numbers contain uppercase characters. In
those instances, we weren't able to look up the docket number using the hidden PACER API because it
doesn't handle uppercase letters on the server (it normalizes them client-side, natch).

With this landed, we can go back to our server and re-run the student work, getting many missed
items.