fix(recap sitemaps): Adds aggressive caching to the sitemaps
This will cache our sitemaps for an entire week. That's a long cache time, but it should be OK since
these don't change much. The next commit changes the sitemaps to be ordered by PK instead of by
dateFiled, which should make these more consistent as the DB changes.

Partially addresses #855.