fix(donations): Ensures that we send the right emails when people are charged
We had this set up wrong previously. What we did was do the charge and then immediately send an
email. This is wrong because we send an email when we get the stripe callback, and therefore we
shouldn't send an email when we make the initial charge.

So, this removes the immediate charge, and adds metadata that the callback can act on. Then in the
callback, it looks for that metadata and sends an appropriate email.