feat(docket page): Add the logic for creating/blocking alert creation
This is pretty annoying code. The difficulty comes from two places:

1. Figuring out if a user can create a docket alert is a combination of both server side information
   (how many alerts do they have in the DB, are they a monthly donor?) and client side information
   (do they have RECAP installed?)

2. There are four different variables that determine whether somebody can create docket alerts on a
   given page:

     - Did they have an alert when the page loaded? If so, let them make it again, full stop.
     - Are they authenticated? If not, don't let them create an alert.
     - Do they get unlimited alerts? If so, let them make an alert.
     - Are they over the limit? If so, no more alerts --- unless of course, they have RECAP
       installed. If that's the case, give them the RECAP bonus. Are they *still* over the limit?

The result of all this is that when the page loads, we handle the first three and a half cases: Did
the page have an alert when it loaded? Is the user authenticated? Do they get unlimited alerts? Are
they over the server-side limit? Then, if we get a message from the RECAP extension saying it's
installed, if the person was previously over the limit, we check if the bonus helps them out. If so,
we unblock alert creation.

The other tricky thing here is political: We're asking for money and we better be nice about it.
Hopefully we've pulled this off too.