feat(docket alerts): Check if the user is a monthly donor
Incredibly, I totally forgot to include this check. Turns out that the logic here is just pretty
complex, with a lot of different pathways to being allowed to make a docket, and somehow I forgot
to add one of the more important ones. This is what comes from working late at night.

We had a monthly donor today that did their donation and then didn't have it work. I set them up
as a free account, and was confused why being a monthly donor didn't work. Well, it's because I
didn't write the code. Anyway, here it is.