feat(registrations): Override the password reset form to send error message
The previous few commits were changing the wrong part of the code, but it sort of works out OK,
because that area of the code should be changed too. The net effect of that code and this code is
that when you try to reset your password or request a confirmation email, and you use an email
address that is not in CourtListener, you get an error message sent to you.

In both cases, the front end will report that everything went swimmingly. This is important so that
people can't probe the system to see who has email addresses on it. But, now, even when things go
swimmingly on the front end, we'll be sure to tell people if there's an issue on the backend.