feat(opinion downloader): Adds user/pass to the opinion downloader as a fallback.
I'm not sure why this is necessary, but I'm hopeful it will help. The issue is that every night when
the downloader runs, a couple thousand PDFs get downloaded fine, but a handful say that they need
to log in. But those ones can't log in because all they have are cookies, not a user/pass, so they
just give up and throw an error. This has been happening for a while, and I don't really know what
causes it, but I'm going to try giving the user/pass as a fallback here and if a re-login is needed,
at least the code can do so.