feat(ia uploader): Convert it *back* into a daemon
There's some wonky stuff in here, but the bottom line is that we want a daemon that runs around the
clock. Once it gets to the end of the collection, it should slow down and wait, occasionally
querying the DB to see if there's new stuff because we've begun the next quarter.

We also want to do all of this in a very performant manner. To that end, we don't use the `count`
method, and we track our progress by just using an iterator. There's some weird stuff in here, but
also lots of comments, so hopefully it makes sense.