feat(twitter card info): Update image attributes (link & height)
According to Facebook's documentation:

https://developers.facebook.com/docs/sharing/webmasters/#images

The height attribute should be:

> Height of image in pixels. Specify height and width for your image to ensure that the image loads
> properly the first time it's shared.

We could do this better by storing the height in our DB, but it seems like Facebook essentially
does that for us after the first load. So this is just an estimate that'll get us started until
Facebook can actually measure and store the info. One less thing for us to do/think about.