refactor(recap doc templates): Merge the regular recap doc template with the new twitter one
Sigh. I belatedly realized that there's no need to have two templates, one for twitter and one for
humans. All that's needed is that whenever the Twitterbot comes crawling, we make sure they get a
page with the right card data. Which page? Well, the regular page is perfectly fine, if we put the
data on it, which we should do anyway.

So this takes the twitter card stuff from the new twitter card template, puts it on the regular
recap document page (something that's good to do anyway), then takes the PDF display and shoves it
into a tab (something we've wanted to, oh, forever). So we get some really nice enhancements to the
regular template and we get rid of the extra Twitter template.