fix(thumbnails): Replace thumbnail pipeline
This took too long to figure out. We had imagemagick in here, but the versions on my laptop and on
the server were creating horrid thumbnails. The version on @jhwak's laptop, meanwhile did a fine job
of it. Anyway, at @jhawk's suggestion, this switches the tool for this job to pdftoppm, which does
a *much* better job, and does so consistently.

But, all is not so easy. Though it does a great job making thumbnails, it has a bit of a weird API:

 - It can't be used with tempfiles because it doesn't accept output file names. Instead it accepts
   a prefix, which it then tacks its own junk onto the end of. But if you leave off the output file
   name, it'll send its output to stdout, which we can capture. So that's what we do instead. Fine.

 - It doesn't let you scale to 'x width with auto scaled height'. Instead, you can only tell it the
   max you want any dimension to be, so we do that instead. Fine. @jhawk thinks using the -r
   resolution flag would be better, but they're similar ideas.

Anyway, this should produce better thumbs than imagemagick.