feat(sitemaps): Cache the sitemap index for one week
I think this should be fine. It'll mean some things take longer to reach Google, but I think that's
OK. The reason this needs to be so high is because grouped facets take a very long time to produce.
Those, alas, are necessary to sitemaps that are faceted by court and page, which greatly improves
performance.