fix(client work): Ensure ordering is not by score
Ordering by score is bad because it means the sort order will be changing under you, especially as
you add content to the database. In order words, documents 2200-2300 when paginating are not the
same when you get to documents 2200-2300 as they were when you started downloading content. The
result of this is that as you get closer to the end of your bulk work, you get more and more
documents in your results that you already crawled, because previously they were earlier in the
sort order. A mess, in short.