fix(client work): Set the page size to be larger than the result size and remove pagination
This is an important tweak. It means that we'll get a *lot* of results at once, but it's important
because otherwise the data will be changing under our search results. We are querying for
is_available:false, which currently returns about 3k items. However, as we proceed, that'll return
fewer and fewer results. When we get about halfway, if we were using a small page size with
pagination, we'd look for items 1500-1600, say, but those wouldn't exist anymore, because only 1500
items would remain that matched is_available:false. That'd be...not ideal. We'd get no results, but
the paginator would probably expect some. It might crash, it might just try the next 100 results,
idk, but it wouldn't be good.