refactor(captions): Update for new data model
This was a tricky one, but the crux of it is that we need a way of sorting the results of the new
Citation model. This could be done by doing a custom manager on the class, but that felt too heavy
and less reliable. Instead, I wrote a new function called sort_keys that has the ordering coded into
it. The basic idea is two things:

1. Try to identify the most important citation for a cluster so that it can go in the caption.

2. For the rest of the citations, try to put them in a logical order.

I've also refactored the caption field here so that it is much smaller.