feat(utils): Add new function for converting a naive date to an aware one in midnight PST
This is a funny thing. We get dates from court websites, but we don't have timezones associated with
them. Solr requires dates with timezone information associated with them. When we pass a naive date
to Solr, it just assumes it's midnight UTC, and converts it as such. This causes the search results
to have UTC dates instead of PST dates, and to be off from the dates in the rest of the site. Users
became confused, and things were bad.

So, here comes the solution. Since we don't actually have times associated with our various dates,
we just assume everything is midnight PST. Then, when we hand it to Solr, things are not naive, and
it does the right conversions. (It converts our PST date to a UTC date in Solr, then when the UTC
date is passed out of Solr and into the template, it's converted to a PST date according to our
TIME_ZONE setting in Django.)