feat(celery): Add a new min_wait param to our celery throttle
This should work better than the min_items param. The idea of the min_items param is that by
limiting the length of the queue, you could slow things down, but that doesn't really work. It just
results in adding and removing things from a very short queue very quickly.

This new solution ensures that whenever maybe_wait is called, you can have a small delay if desired.