Start automatically running tests on travis-ci.com
Motivation: there's no CI/CD for this project right now.
Humans forget to run the test suite before merging.
This can lead to bugs getting deployed to production.
Let's use travis-ci.com, a CI/CD platform that's free for
open source projects, to run tests. We'll start small with
just testing one module, `cl.corpus_importer`.

There are lots of dependencies that need to be
installed before the tests in this repo can be run.
Since we're already using ansible to deploy this project,
let's try using ansible to setup test dependencies.
Hopefully, in the future the ansible scripts to test and deploy
can be consolidated in some sensible way. For now, let's just
duplicate some logic in https://github.com/freelawproject/freelawmachine.
In the future we can refactor ansible scripts in both these repos to be
more maintainable and less hacky for both dev environments and CI/CD.

For now we're just running a small set of tests to prove this works.
I'm not sure if the test files are structured according to Django
best practices.