fix(dependencies): Adjust psychopg to not install binaries
I guess this is what's needed, according to the documentation and a issue they have filed. It feels
dumb, but we get a warning now and this seems to be the way to address it. When 2.8 arrives (it's in
dev now), we can upgrade to that and it seems like it will solve the remaining problems, I guess.