feat(browser tools): Remove open search install page
It turns out it wasn't linked to from anywhere, and it hadn't received a single visit this year.
Also, the JS function for doing this, `AddSearchProvider` has been deprecated from Chrome and is
getting deprecated from Firefox. Tis a dying thing.

We do keep the opensearch.xml file that is in base.html in a <link> element.