fix(ia uploader): Remove confusing method call to mark_ia_upload_needed
This is a very odd bug. After we did our upload to IA, we were coming back around and trying to
mark the docket as needing upload a second time. Luckily, this did nothing because if a docket is
already marked as needing upload, it doesn't do anything. What we should have been doing is marking
the docket as *not* needing upload.

We got away with this because of the way we filter dockets for needing upload. They need to have a
couple different columns set before they get uploaded, which this didn't quite accomplish even in
it's bugginess.

I don't like this commit because it's fixing something that ain't broke. There's no symptoms of this
causing issues. I just noticed that it looked weird when fixing a different bug. It looks like a
no-op anyway in this case, so I think we're fine, but it's surely pretty weird and distressing.