refactor(docket alerts): Reduce number of tasks created by crawled PACER RSS feeds
Prior to this commit, we create one task per docket within an RSS feed. This creates an insane
number of tasks which can take a LOT of processing. A better way to handle this is to create one
alert-sending task per RSS feed that we process, and to do so as part of the original processing
chain. This commit accomplishes that change in a minimally invasive kind of way.