feat(fjc db migrations): Add index together for court and docket number
The latest FJC data lacks the JURY field for some reason, making it impossible to do a diff from
their old data against the new (every line has changed due to the missing field). If we could do
that, then we'd be in the happy place of just importing changed lines. Woo!

But we're not in the happy place, so we need a better strategy that can update all the rows. This
code provides a multicolumn index so we can do lookups faster when merging new data.