feat(corpus tasks): New task to filter based on tag names
This will help our client work where we're resuming a download that didn't work completely. In this
case, the completed items are tagged and the remainder are missing. So we'll look up items in the
IDB, and then we'll know if we already downloaded them based on this filter. If we get hits, we
abort. If not, we continue.