fix(free opinion scraper): Add pacer_seq_no to the scraper
Note that this uses pacer_seq_no rather than pacer_sequence_number. pacer_seq_no is what we use
in Juriscraper, where things are supposed to be closer to PACER standards and values, while
pacer_sequence_number is supposed to be our normalized values. All the other instances of this
field in CL are pacer_sequence_number, not pacer_seq_no. Why the difference?

The reason is because this model, as part of the scraper, is more like Juriscraper. It holds values
temporarily until they can be processed by the PDF downloader. The values here are more like raw
values, waiting to be ingested than they are like normalized values.