fix(models): Add automatic deletion code for validation errors on RECAPDocument objects
This code doesn't thrill me and is likely to cause some confusion at some point, but the main idea
is sound, I think. Basically, right now we have some items in the DB that violate uniqueness
constraints that we'd like to have. Currently when we try to save these items, they just crash, and
crash, and crash, every time we save them. This change makes it so instead of crashing, these items
get cleaned up by deleting the item that's not being saved. In some cases that will be the "wrong"
item, but the idea is that it's better to delete the wrong item than to have two items in the DB. By
at least deleting *one* of them, we can focus on the other and make it as good as possible. Gotta
start somewhere.