fix(rss scraper): Flip the logic of processing cut off
The goal of this code block was to make sure we didn't crawl a court too much. If we just started
crawling a few minutes ago, we shouldn't try to do so again. Unfortunately, it looks like the logic
is backwards here, making it do the opposite thing: If the court hasn't been crawled in a long time,
then don't crawl it again now.

This shouldn't have a major effect because courts don't usually take so long that this code would
get triggered (they have to be processing *longer* than the PROCESSING_IN_PROGRESS variable). But,
if that ever happens, you wind up in a place where you'll never crawl the court again, since it'll
always get cut off by this block.

By flipping the logic, we have the desired intent:

1. If a court is processing for too long, we assume it crashed and needs to be tried again.
2. If a court is still processing for a short period of time, we ignore it and come back later.

Addresses #934