feat(tests): Allow DB queries in SimpleTestCase
In theory these test cases shouldn't hit the DB, but since we're upgrading from 1.8 to 1.9, which
imposes this requirement, we're going to bypass it rather than rewrite these tests. If you ever
are reading this message, you should probably go see if these lines of code can be removed via a
simple refactor. Now is not the time, though.