refactor(tasks): Consolidate solr update tasks to a single versatile task
Now instead of having tasks for updating people, and audio, and dockets, and this and that, we have
one task that can handle them all.

This was triggered by our update to Celery, which now uses the JSON renderer instead of pickle. That
change means that we can no longer pass django model instances around because they can't be
serialized in JSON, so I got to work fixing that. Along the way, this mega refactor became
necessary.

More of this refactor will come in the next two commits.