fix(search): Fix data corruption in solr across recap documents within a docket
When we index an entire docket all at once, we only grab the docket values once and we only grab the
docket entry values as few times as needed. Before this fix, as we looped, we ran the risk of
applying the metadata from an earlier iteration through the loop to the next iteration through it.

For example, if we had a docket entry that lacked an `entry_number` value after one that had one,
the latter one could easily get the value from the earlier one, since that value was just carried
across iterations in the loop and wasn't nuked between iterations.

This change ensures that we don't mess up our caches of these values as we iterate through the
nested loops.