fix(citations): Updates task and its callers to not pass objects
Needed for latest celery version, which uses JSON instead of pickle for task serialization. This
also takes the liberty of updating *both* the cl_find_citation command *and* the tasks themselves
to better mirror the style of the update_index commands. This simplified the code a bit and made it
more similar across tasks.