fix(api filters): Upgrade for filters 1.0
Two interesting tweaks here. First, the `action` parameter became the method parameter. They seem to
have had their reasons for this.

Second, the name parameter became the field_name parameter. This one surprised me a bit because
along with tweaking the parameter value, I also needed to tweak the values that it got, so it wasn't
just a straight tweak to the parameter name.

Why did the values need to get tweaked? I'm not sure. The old ones had tests that seemed to work
previously, but when they were run now, the system (rightly) informed us that they didn't point to
valid fields. So were they ignored in the past? No, because the tests passed. Did they work via some
magic? Maybe? I'm not sure, but I got them working again, they seem to be correctly configured in a
way the system works and which makes the tests pass, so I think we're in business. Still though,
curious stuff, and I don't particularly love it.