fix(apache): Make Apache perform
This godawful configuration was ensuring that two python processes served all of our requests for
the past many years. JFC, ever wonder why giving Apache more power via the worker module isn't
working, well, guess what, this is why. Somehow it's in the site configuration. So that's effed.

Prior to this our performance hit a cliff. After this, things fly. Damn it.